move mitmproxy launching into separate file
TODO incorporate
